requirements.txt
kfp==1.8.6
google-cloud-aiplatform==1.6.0
google-cloud-pipeline-components==0.1.7